Create separate file for bq util functions. Setup config variables for enablingBigQuery
